Finish JPA CRUD implementation

+++ src/main/java/com/mashanlote/model/weather/UpdateWeatherObservationRequest.java
@@ -4,6 +4,6 @@ import java.util.UUID;
 
 public record UpdateWeatherObservationRequest(
         UUID weatherObservationId,
-        Integer temperature,
+        Double temperature,
         Integer weatherTypeId
 ) { }
